LibC: Move `dlfcn_integration.h` to the `bits` directory
